+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 * Add a settings to wait for unity video preparation before playing
 * Add hooks to prevent custom savefile stream buffering
 * Add hook to SDL_CreateWindowWithProperties()
+* Implement SDL_OpenAudioDeviceStream()
 
 ### Changed